requirements.txt
transformers==4.57.3
datasets==4.4.1
tokenizers==0.22.1
accelerate==1.12.0
httpx>=0.28.1
httpcore==1.0.9
torch==2.9.1
perft==0.5.0
pandas
numpy
scikit-learn
matplotlib
streamlit
sentencepiece
python-docx
pdfplumber

